requirements.txt
Flask==3.0.0
langchain>=0.1.0
langchain-google-genai>=1.0.0
sentence-transformers==2.2.2
faiss-cpu>=1.9.0
pypdf>=6.0.0
python-docx>=1.1.0
google-generativeai>=0.3.0
python-dotenv==1.0.0
